+++ Package.swift
@@ -30,6 +30,7 @@ let package = Package(
                 .swiftLanguageMode(.v5)
             ],
             linkerSettings: [
+                .linkedFramework("Carbon"),
                 .linkedFramework("CoreServices"),
                 .linkedFramework("InputMethodKit"),
                 .linkedFramework("WebKit")