+++ Sources/MyIMEMacOS/SecureInputDetector.swift
@@ -1,0 +1,7 @@
+import Carbon
+
+enum SecureInputDetector {
+    static var isEnabled: Bool {
+        IsSecureEventInputEnabled()
+    }
+}